Extract pgn logic to its own header. Also fix a bug I introduced changing the wrong 9 to an 8.
